refactor(employees): move status validation helper from lib/helpers (CP-5)
ensureEmployeeActive, ensureEmployeeNotTerminated and ensureEmployeeNotOnVacation
now live at src/modules/employees/status.ts, next to the employees module
they actually validate. Previously they sat in src/lib/helpers/ — a catch-all
that only held this one file.

9 occurrence services (absences, accidents, cpf-analyses, labor-lawsuits,
medical-certificates, ppe-deliveries, promotions, vacations, warnings)
updated to import from the new path. src/modules/occurrences/CLAUDE.md
reflects the new location. src/lib/helpers/ directory removed (was empty
after the move).

Behavior preserved — same functions, same queries, same throws.
